refactor(vocabularies): split 'superdesk-vocabularies' module.
`superdesk-vocabularies` becomes:
```
.
├── controllers
│   ├── VocabularyConfigController.js
│   ├── VocabularyEditController.js
│   └── index.js
├── directives
│   ├── VocabularyConfigDirective.js
│   ├── VocabularyConfigModal.js
│   └── index.js
├── services
│   ├── SchemaFactory.js
│   ├── VocabularyService.js
│   └── index.js
├── styles
│   └── vocabularies.less
├── tests
│   └── vocabularies.spec.js
├── views
│   ├── settings.html
│   ├── vocabulary-config-modal.html
│   └── vocabulary-config.html
└── index.js

6 directories, 15 files
```
