+++ scripts/superdesk-vocabularies/controllers/index.js
@@ -1,0 +1,2 @@
+export { VocabularyConfigController } from './VocabularyConfigController';
+export { VocabularyEditController } from './VocabularyEditController';
