+++ scripts/superdesk-vocabularies/directives/VocabularyConfigDirective.js
@@ -1,0 +1,9 @@
+export function VocabularyConfigDirective() {
+    return {
+        scope: {
+            vocabulary: '=',
+        },
+        controller: 'VocabularyConfig',
+        templateUrl: 'scripts/superdesk-vocabularies/views/vocabulary-config.html'
+    };
+}
