+++ scripts/superdesk-vocabularies/directives/VocabularyConfigModal.js
@@ -1,0 +1,9 @@
+export function VocabularyConfigModal() {
+    return {
+        scope: {
+            vocabulary: '=',
+        },
+        controller: 'VocabularyEdit',
+        templateUrl: 'scripts/superdesk-vocabularies/views/vocabulary-config-modal.html'
+    };
+}
