+++ scripts/superdesk-vocabularies/directives/index.js
@@ -1,0 +1,2 @@
+export { VocabularyConfigDirective } from './VocabularyConfigDirective';
+export { VocabularyConfigModal } from './VocabularyConfigModal';
