+++ scripts/superdesk-vocabularies/services/index.js
@@ -1,0 +1,2 @@
+export { VocabularyService } from './VocabularyService';
+export { SchemaFactory } from './SchemaFactory';
